refactor(shared): move auth utils
