+++ frontend/packages/frontend/src/pages/auth/LogoutPage.tsx
@@ -1,6 +1,6 @@
 import {useEffect} from 'react';
 import {useNavigate} from 'react-router-dom';
-import { clearAuthToken } from '@photobank/shared/api/auth';
+import { clearAuthToken } from '@photobank/shared/auth';
 import { loggingOutMsg } from '@photobank/shared/constants';
 
 export default function LogoutPage() {
