+++ frontend/packages/shared/src/auth/index.ts
